Fix CodeQL alerts in the Tecplot ASCII and SU2 binary mesh writers
Use unsigned long for the field counter of the Tecplot ASCII writer, which
is compared with the size of the field names, and rename the offset of the
boundary elements of the SU2 binary mesh writer, which hid the one of the
volume elements.

+++ SU2_CFD/src/output/filewriter/CTecplotFileWriter.cpp
@@ -43,7 +43,7 @@ void CTecplotFileWriter::WriteData(string val_filename){
 
   const vector<string> fieldNames = dataSorter->GetFieldNames();
 
-  unsigned short iVar;
+  unsigned long iVar;
 
   unsigned long iPoint, iElem;
